Rationalize the localizations module
Extract translations to another file to keep the dependencies sane.

JIRA: RAIL-3383
